Separates the Import and Update Process
This patch separates the import and update processes and adds
examples for use.

Import

The import process largely remains the same with no major changes
to the logic other than to remove the updating of cards.

Update

Whereas the import logic retrieves data from sources, the update
logic retrieves data for cards in the board from matching artifacts
from launchpad or bugzilla.  Blueprints and StoryBoard support are
TBD.  The update logic may change, but the contract of calling
board_manager.update should not.